backend/generate_vps/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
playwright==1.40.0
requests==2.31.0
psutil==5.9.6
dnspython==2.8.0
pytz==2023.3
stringcase==1.2.0
